requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-mysql==3.9
django-cors-headers==4.3.1
pika==1.1.0
psycopg2==2.9.9